pyproject.toml
[project]
name = "quant-execution-engine"
version = "0.2.0"
requires-python = ">=3.10, <3.13"

dependencies = [
  "longport>=0.2.77",
  "pyyaml",
]

[dependency-groups]
dev = [
  "pytest>=8",
  "pytest-cov",
  "ruff>=0.5",
  "mypy>=1.10",
]

[project.optional-dependencies]
cli = ["rich>=13"]

[project.scripts]
stockq = "quant_execution_engine.cli:app"
qexec = "quant_execution_engine.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["quant_execution_engine*"]

[build-system]
requires = ["setuptools>=77.0.3"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
# Select rule sets to enable.
# "E", "F" correspond to Flake8's native rules.
# "I" corresponds to isort.
# "UP" corresponds to pyupgrade.
select = ["E", "F", "I", "UP", "B"] # B is flake8-bugbear

# Or, enable all rules and then ignore the ones you don't need
# select = ["ALL"]
# ignore = ["D100", "D101"] # Ignore specific rules

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.coverage.run]
source = ["src/quant_execution_engine"]

[tool.coverage.report]
show_missing = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=quant_execution_engine --cov-report=term-missing --cov-fail-under=75 -m 'not integration and not e2e and not slow'"
markers = [
  "unit: fast, isolated tests",
  "integration: tests that touch external APIs or DB",
  "e2e: end-to-end workflow tests",
  "slow: long-running tests",
  "requires_api: needs external API access",
  "requires_db: needs database access",
]
filterwarnings = [
  "error::DeprecationWarning",
  "ignore:.*distutils.*:DeprecationWarning",
]
